+++ src/ops/cloud/VariablesOps.ts
@@ -100,7 +100,7 @@ export async function listVariables(
       return false;
     }
     //Delete variables from full export so they aren't mistakenly used for determining usage
-    delete fullExport.global.variables;
+    delete fullExport.global.variable;
     headers.push('Used'['brightCyan']);
   }
   const table = createTable(headers);
@@ -324,11 +324,6 @@ export async function describeVariable(
   usage = false,
   json = false
 ): Promise<boolean> {
-  const spinnerId = createProgressIndicator(
-    'indeterminate',
-    0,
-    `Describing variable ${variableId}...`
-  );
   try {
     const variable = (await readVariable(variableId)) as VariableSkeleton & {
       locations: string[];
@@ -337,23 +332,13 @@ export async function describeVariable(
       try {
         const fullExport = await getFullExportConfig(file);
         //Delete variables from full export so they aren't mistakenly used for determining usage
-        delete fullExport.global.variables;
+        delete fullExport.global.variable;
         variable.locations = getIdLocations(fullExport, variableId, true);
       } catch (error) {
-        stopProgressIndicator(
-          spinnerId,
-          `Error determining usage for variable with id ${variableId}`,
-          'fail'
-        );
         printError(error);
         return false;
       }
     }
-    stopProgressIndicator(
-      spinnerId,
-      `Successfully retrieved variable ${variableId}`,
-      'success'
-    );
     if (json) {
       printMessage(variable, 'data');
     } else {
@@ -404,11 +389,6 @@ export async function describeVariable(
     }
     return true;
   } catch (error) {
-    stopProgressIndicator(
-      spinnerId,
-      `Error describing variable ${variableId}`,
-      'fail'
-    );
     printError(error);
   }
   return false;
@@ -639,7 +619,7 @@ export async function importVariablesFromFiles(): Promise<boolean> {
       try {
         const data = fs.readFileSync(file, 'utf8');
         const fileData: VariablesExportInterface = JSON.parse(data);
-        const count = Object.keys(fileData.variables).length;
+        const count = Object.keys(fileData.variable).length;
         total += count;
         await importVariables(fileData);
         updateProgressIndicator(